requirements.txt
fitz==0.0.1.dev2
joblib==1.4.2
langchain==0.3.26
langchain_community==0.3.27
langchain_core==0.3.68
langchain_groq==0.3.6
langchain_huggingface==0.3.0
langchain_text_splitters==0.3.8
matplotlib==3.8.0
nltk==3.8.1
numpy==2.3.1
pandas==2.3.1
plotly==5.9.0
python-dotenv==1.1.1
streamlit==1.30.0
tensorflow==2.19.0
tensorflow_intel==2.17.0
yfinance==0.2.63